+++ src/constants.ts
@@ -1,4 +1,7 @@
-import { RelationshipClass } from '@jupiterone/integration-sdk-core';
+import {
+  RelationshipClass,
+  RelationshipDirection,
+} from '@jupiterone/integration-sdk-core';
 
 export const SetDataKeys = {
   ACCOUNT_ENTITY: 'ACCOUNT_ENTITY',
@@ -7,19 +10,10 @@ export const SetDataKeys = {
 export const StepIds = {
   FETCH_ACCOUNT: 'fetch-account',
   FETCH_FINDINGS: 'fetch-findings',
+  BUILD_FINDING_CVE_CWE_RELATIONSHIPS: 'build-issue-cve-cwe-relationships',
 };
 
 export const Entities = {
-  CVE: {
-    _type: 'cve',
-    _class: 'Vulnerability',
-    resourceName: 'CVE',
-  },
-  CWE: {
-    _type: 'cwe',
-    _class: 'Weakness',
-    resourceName: 'CWE',
-  },
   SNYK_ACCOUNT: {
     _type: 'snyk_account',
     _class: ['Service', 'Account'],
@@ -32,23 +26,41 @@ export const Entities = {
   },
 };
 
+export const TargetEntities = {
+  CVE: {
+    _type: 'cve',
+    _class: 'Vulnerability',
+    resourceName: 'CVE',
+  },
+  CWE: {
+    _type: 'cwe',
+    _class: 'Weakness',
+    resourceName: 'CWE',
+  },
+};
+
 export const Relationships = {
-  FINDING_IS_CVE: {
-    _type: 'snyk_finding_is_cve',
-    sourceType: Entities.SNYK_FINDING._type,
-    _class: RelationshipClass.IS,
-    targetType: Entities.CVE._type,
-  },
-  FINDING_EXPLOITS_CWE: {
-    _type: 'snyk_finding_exploits_cwe',
-    sourceType: Entities.SNYK_FINDING._type,
-    _class: RelationshipClass.EXPLOITS,
-    targetType: Entities.CWE._type,
-  },
   SERVICE_IDENTIFIED_FINDING: {
     _type: 'snyk_service_identified_snyk_finding',
     sourceType: Entities.SNYK_ACCOUNT._type,
     _class: RelationshipClass.IDENTIFIED,
     targetType: Entities.SNYK_FINDING._type,
+  },
+};
+
+export const MappedRelationships = {
+  FINDING_IS_CVE: {
+    _type: 'snyk_finding_is_cve',
+    sourceType: Entities.SNYK_FINDING._type,
+    _class: RelationshipClass.IS,
+    targetType: TargetEntities.CVE._type,
+    direction: RelationshipDirection.FORWARD,
+  },
+  FINDING_EXPLOITS_CWE: {
+    _type: 'snyk_finding_exploits_cwe',
+    sourceType: Entities.SNYK_FINDING._type,
+    _class: RelationshipClass.EXPLOITS,
+    targetType: TargetEntities.CWE._type,
+    direction: RelationshipDirection.FORWARD,
   },
 };