+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.1.0] - 2021-08-19
+
 ### Changed
 
 - Use the `/aggregated-issues` endpoint to list issues, which is recommended by